Moved server related classes in seperate package.
