+++ ContextDesktop/src/main/java/com/github/migi_1/Context/server/AccelerometerMessageHandler.java
@@ -1,5 +1,6 @@
-package com.github.migi_1.Context;
+package com.github.migi_1.Context.server;
 
+import com.github.migi_1.Context.Main;
 import com.github.migi_1.ContextMessages.AccelerometerMessage;
 import com.github.migi_1.ContextMessages.MessageListener;
