+++ ContextDesktop/src/main/java/com/github/migi_1/Context/server/ActiveServerState.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.Context;
+package com.github.migi_1.Context.server;
 
 import com.jme3.network.Server;
