+++ ContextDesktop/src/main/java/com/github/migi_1/Context/server/ClientFinder.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.Context;
+package com.github.migi_1.Context.server;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
