+++ ContextDesktop/src/test/java/com/github/migi_1/Context/TestInactiveServerState.java
@@ -6,6 +6,8 @@ import static org.junit.Assert.assertNotEquals;
 import org.junit.Test;
 import org.mockito.Mockito;
 
+import com.github.migi_1.Context.server.ActiveServerState;
+import com.github.migi_1.Context.server.InactiveServerState;
 import com.jme3.network.Server;
 
 /**
