+++ ContextDesktop/src/test/java/com/github/migi_1/Context/TestServerWrapper.java
@@ -12,6 +12,7 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
+import com.github.migi_1.Context.server.ServerWrapper;
 import com.jme3.network.Network;
 import com.jme3.network.Server;
